+++ packages/orq-rc/RELEASES.md
@@ -659,3 +659,13 @@ Based on:
 - [python v3.2.0-rc.28] packages/orq-rc
 ### Releases
 - [PyPI v3.2.0-rc.28] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.28 - packages/orq-rc
+
+## 2025-03-06 10:08:39
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.510.0 (2.541.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.2.0-rc.29] packages/orq-rc
+### Releases
+- [PyPI v3.2.0-rc.29] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.29 - packages/orq-rc